PO-2254 fixed conflict, split IT case into 2 classes
